+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 0.1.1
 
-* Launchurl with browser.
-* Generate new random contract name.
-* Create new wallet.
+* New utility to Launchurl with browser.
+* New utility to Generate new random contract name.
+* Better handler to create new wallet.
 * Bound existing/new Wallet.
 
 ## 0.1.0